Fixed vite config.

+++ vite.config.ts
@@ -15,17 +15,20 @@ export default defineConfig({
       name: 'rus-dist-cleanup',
       closeBundle: isP
         ? async () => {
-            // Remove hash from assets
-            const fs = require('fs');
-            const path = require('path');
-            const dir = path.resolve(__dirname, 'dist/assets');
-            const files = fs.readdirSync(dir);
-            files.forEach((file: string) => {
-              // Remove hash from file name
-              const newFile = file.replace(/-[a-z0-9]{8}\./, '.');
-              fs.renameSync(path.resolve(dir, file), path.resolve(dir, newFile));
-            });
+          // Remove hash from assets
+          const fs = require('fs');
+          const path = require('path');
+          const dir = path.resolve(__dirname, 'dist/assets');
+          if (!fs.existsSync(dir)) {
+            return;
           }
+          const files = fs.readdirSync(dir);
+          files.forEach((file: string) => {
+            // Remove hash from file name
+            const newFile = file.replace(/-[a-z0-9]{8}\./, '.');
+            fs.renameSync(path.resolve(dir, file), path.resolve(dir, newFile));
+          });
+        }
         : undefined,
     },
   ],
@@ -37,30 +40,30 @@ export default defineConfig({
   ...(isP
     ? {}
     : {
-        build: {
-          outDir: 'dist',
-          emptyOutDir: true,
-          minify: true,
-          watch: {
-            include: 'src/**',
-            exclude: 'node_modules/**, .git/**, dist/**, .vscode/**',
-          },
-          assetsDir: 'assets',
-          cssCodeSplit: false,
-          sourcemap: false,
-          ssr: false,
-          rollupOptions: {
-            treeshake: true,
-            output: {
-              entryFileNames: `assets/index.js`,
-              chunkFileNames: `assets/index-chunk.js`,
-              assetFileNames: `assets/[name].[ext]`,
-            },
-          },
-          // lib: {
-          //   entry: 'src/main.ts',
-          //   name: 'vue3-vite-example',
-          // }
+      build: {
+        outDir: 'dist',
+        emptyOutDir: true,
+        minify: true,
+        watch: {
+          include: 'src/**',
+          exclude: 'node_modules/**, .git/**, dist/**, .vscode/**',
         },
-      }),
+        assetsDir: './assets',
+        cssCodeSplit: false,
+        sourcemap: false,
+        ssr: false,
+        rollupOptions: {
+          treeshake: true,
+          output: {
+            entryFileNames: `assets/index.js`,
+            chunkFileNames: `assets/index-chunk.js`,
+            assetFileNames: `assets/[name].[ext]`,
+          },
+        },
+        // lib: {
+        //   entry: 'src/main.ts',
+        //   name: 'vue3-vite-example',
+        // }
+      },
+    }),
 });
